module exports and tests fixed
