+++ test/api/utils.js
@@ -15,6 +15,55 @@
 import utils from '../../src/api/utils';
 
 describe('api utils module', () => {
+	const port = 7000;
+	const localNode = 'localhost';
+	const externalNode = 'external';
+	const sslNode = 'sslPeer';
+	const externalTestnetNode = 'testnet';
+	const GET = 'GET';
+	const POST = 'POST';
+	const defaultMethod = POST;
+	const defaultEndpoint = 'transactions';
+	const defaultPeers = [localNode, externalNode];
+	const defaultSSLPeers = [localNode, externalNode, sslNode];
+	const defaultTestnetPeers = [localNode, externalTestnetNode];
+
+	let LSK;
+	let sendRequestResult;
+	let sendRequestStub;
+
+	beforeEach(() => {
+		LSK = {
+			randomPeer: false,
+			currentPeer: localNode,
+			defaultPeers: [].concat(defaultPeers),
+			defaultSSLPeers: [].concat(defaultSSLPeers),
+			defaultTestnetPeers: [].concat(defaultTestnetPeers),
+			bannedPeers: [],
+			port,
+			options: {
+				node: localNode,
+			},
+			nethash: {
+				foo: 'bar',
+			},
+
+			parseOfflineRequests: () => ({
+				requestMethod: GET,
+			}),
+			setTestnet: () => {},
+			setNode: () => {},
+			sendRequest: () => {},
+		};
+		sendRequestResult = { success: true, sendRequest: true };
+		sendRequestStub = sinon.stub(LSK, 'sendRequest').resolves(Object.assign({}, sendRequestResult));
+	});
+
+	afterEach(() => {
+		sendRequestStub.restore();
+	});
+
+
 	describe('exports', () => {
 		it('should be an object', () => {
 			(utils).should.be.type('object');
@@ -76,6 +125,213 @@ describe('api utils module', () => {
 				'key:/;?': 'value:/;?',
 			});
 			(queryString).should.be.equal('key%3A%2F%3B%3F=value%3A%2F%3B%3F');
+		});
+	});
+
+	describe('#checkOptions', () => {
+		const { checkOptions } = utils;
+		const goodOptions = {
+			key1: 'value 1',
+			key2: 2,
+		};
+
+		it('should throw an error if any option is undefined', () => {
+			const optionsWithUndefined = Object.assign({
+				badKey: undefined,
+			}, goodOptions);
+
+			(checkOptions.bind(null, optionsWithUndefined)).should.throw('"badKey" option should not be undefined');
+		});
+
+		it('should throw an error if any option is NaN', () => {
+			const optionsWithNaN = Object.assign({
+				badKey: NaN,
+			}, goodOptions);
+
+			(checkOptions.bind(null, optionsWithNaN)).should.throw('"badKey" option should not be NaN');
+		});
+
+		it('should return the options if they are all ok', () => {
+			const result = checkOptions(Object.assign({}, goodOptions));
+			(result).should.be.eql(goodOptions);
+		});
+	});
+
+	describe('#serialiseHTTPData', () => {
+		const { serialiseHTTPData } = utils;
+		const queryStringData = 'key%2F1=value%20%252&key3=4';
+
+		let data;
+		let serialisedData;
+
+		beforeEach(() => {
+			data = {
+				' key/1 ': '  value %2',
+				key3: 4,
+			};
+			serialisedData = serialiseHTTPData(data);
+		});
+
+		it('should prepend a question mark to the query string', () => {
+			(serialisedData).should.equal(`?${queryStringData}`);
+		});
+	});
+
+	describe('#wrapSendRequest', () => {
+		const { wrapSendRequest } = utils;
+		const value = '123';
+
+		let options;
+		let getDataFnResult;
+		let getDataFnStub;
+		let constructRequestDataResult;
+		let constructRequestDataStub;
+		let restoreConstructRequestDataStub;
+		let callback;
+		let returnedFunction;
+
+		beforeEach(() => {
+			options = {
+				key1: 'value 1',
+				key2: 2,
+			};
+			getDataFnResult = {
+				key3: 'value3',
+				key4: 4,
+			};
+			constructRequestDataResult = {
+				key5: 'value 5',
+				key6: 6,
+			};
+			getDataFnStub = sinon.stub().returns(Object.assign({}, getDataFnResult));
+			constructRequestDataStub = sinon.stub()
+				.returns(Object.assign({}, constructRequestDataResult));
+			// eslint-disable-next-line no-underscore-dangle
+			restoreConstructRequestDataStub = utils.__set__('constructRequestData', constructRequestDataStub);
+			returnedFunction = wrapSendRequest(defaultMethod, defaultEndpoint, getDataFnStub);
+			callback = () => {};
+		});
+
+		afterEach(() => {
+			restoreConstructRequestDataStub();
+		});
+
+		it('should return a function', () => {
+			(returnedFunction).should.be.type('function');
+		});
+
+		describe('returned function', () => {
+			it('should call the provided getData function on the provided value and options', () => {
+				return returnedFunction.call(LSK, value, options)
+					.then(() => {
+						(getDataFnStub.calledWithExactly(value, options)).should.be.true();
+					});
+			});
+
+			it('should construct request data using the provided data and options', () => {
+				return returnedFunction.call(LSK, value, options)
+					.then(() => {
+						(constructRequestDataStub.calledWithExactly(getDataFnResult, options)).should.be.true();
+					});
+			});
+
+			it('should send a request with the constructed data and a callback if options are provided', () => {
+				return returnedFunction.call(LSK, value, options, callback)
+					.then(() => {
+						(sendRequestStub.calledWithExactly(
+							defaultMethod, defaultEndpoint, constructRequestDataResult, callback,
+						)).should.be.true();
+					});
+			});
+
+			it('should send a request with the constructed data and a callback if options are not provided', () => {
+				return returnedFunction.call(LSK, value, callback)
+					.then(() => {
+						(sendRequestStub.calledWithExactly(
+							defaultMethod, defaultEndpoint, constructRequestDataResult, callback,
+						)).should.be.true();
+					});
+			});
+
+			it('should return the result of the sent request', () => {
+				return returnedFunction.call(LSK, value, callback)
+					.then((result) => {
+						(result).should.be.eql(sendRequestResult);
+					});
+			});
+		});
+	});
+
+	describe('#constructRequestData', () => {
+		const address = '18160565574430594874L';
+		const customAddress = '123l';
+		const defaultRequestLimit = 10;
+		const defaultRequestOffset = 101;
+		const optionsObject = {
+			limit: defaultRequestLimit,
+			offset: defaultRequestOffset,
+		};
+		const expectedObject = {
+			address,
+			limit: defaultRequestLimit,
+			offset: defaultRequestOffset,
+		};
+		const optionsWithConflictObject = {
+			address: customAddress,
+			limit: 4,
+			offset: 5,
+		};
+		const resolvedConflictObject = {
+			address: customAddress,
+			limit: defaultRequestLimit,
+			offset: defaultRequestOffset,
+		};
+
+		it('should merge a data object with an options object', () => {
+			const requestData = utils.constructRequestData({ address }, optionsObject);
+			(requestData).should.be.eql(expectedObject);
+		});
+
+		it('should recognise when a callback function is passed instead of an options object', () => {
+			const providedObj = { address };
+			const requestData = utils.constructRequestData(providedObj, () => true);
+			(requestData).should.be.eql(providedObj);
+		});
+
+		it('should prioritise values from the data object when the data object and options object conflict', () => {
+			const requestData = utils.constructRequestData(
+				{ limit: defaultRequestLimit, offset: defaultRequestOffset }, optionsWithConflictObject,
+			);
+			(requestData).should.be.eql(resolvedConflictObject);
+		});
+	});
+
+	describe('#optionallyCallCallback', () => {
+		const { optionallyCallCallback } = utils;
+		const result = 'result';
+		const spy = sinon.spy();
+
+		it('should return the result with a callback', () => {
+			const returnValue = optionallyCallCallback(spy, result);
+			(returnValue).should.equal(result);
+		});
+
+		it('should return the result without a callback', () => {
+			const returnValue = optionallyCallCallback(undefined, result);
+			(returnValue).should.equal(result);
+		});
+
+		it('should not call the callback if it is not a function', () => {
+			(optionallyCallCallback.bind(null, { foo: 'bar' }, result)).should.not.throw();
+		});
+
+		it('should not call the callback if it is undefined', () => {
+			(optionallyCallCallback.bind(null, undefined, result)).should.not.throw();
+		});
+
+		it('should call the callback with the result if callback is a function', () => {
+			optionallyCallCallback(spy, result);
+			(spy.calledWithExactly(result)).should.be.true();
 		});
 	});
 });
